feat: update quantity type to Double in repositories and add pedido/ordenCompra fields in DTOs and mappers

+++ src/main/java/com/coagronet/articuloKardex/repositories/ArticuloKardexRepository.java
@@ -20,7 +20,7 @@ public interface ArticuloKardexRepository extends JpaRepository<ArticuloKardex, 
 
 		Long getPresentacionId();
 
-		BigDecimal getCantidad();
+		Double getCantidad();
 
 	}
